Fixing directory structure
