+++ src/snooker_ball_tracker/models/settings/__init__.py
@@ -1,0 +1,3 @@
+from .ball_detection_tab_model import BallDetectionTabModel
+from .colour_detection_tab_model import ColourDetectionTabModel
+from .ball_detection_setting_model import BallDetectionSettingModel
