+++ src/snooker_ball_tracker/views/components/__init__.py
@@ -1,0 +1,6 @@
+from .combobox import Ui_Combobox
+from .label import Ui_Label
+from .line import Ui_Line
+from .pushbutton import Ui_PushButton
+from .radiobutton import Ui_RadioButton
+from .slider import Ui_Slider
