+++ src/snooker_ball_tracker/views/logging/__init__.py
@@ -1,0 +1,2 @@
+from .balls_potted import Ui_BallsPotted
+from .snapshot_info import Ui_SnapshotInfo
