+++ src/snooker_ball_tracker/views/logging_view.py
@@ -1,13 +1,10 @@
 import PyQt5.QtCore as QtCore
 import PyQt5.QtGui as QtGui
 import PyQt5.QtWidgets as QtWidgets
+from snooker_ball_tracker.models import LoggingModel
 
-from .components.label import Ui_Label
-from .components.line import Ui_Line
-from .components.pushbutton import Ui_PushButton
-from .logging_balls_potted import Ui_BallsPotted
-from .logging_snapshot_info import Ui_SnapshotInfo
-from snooker_ball_tracker.models.logging_model import LoggingModel
+from .components import Ui_Label, Ui_Line, Ui_PushButton
+from .logging import Ui_BallsPotted, Ui_SnapshotInfo
 
 
 class LoggingView(QtWidgets.QGroupBox):
