+++ src/snooker_ball_tracker/views/settings/colour_detection_tab_view.py
@@ -4,16 +4,11 @@ import PyQt5.QtCore as QtCore
 import PyQt5.QtGui as QtGui
 import PyQt5.QtWidgets as QtWidgets
 import snooker_ball_tracker.settings as s
-from snooker_ball_tracker.models.observer import Observer
-from snooker_ball_tracker.models.colour_detection_model import \
-    ColourDetectionTabModel
+from snooker_ball_tracker.models import Observer
+from snooker_ball_tracker.models.settings import ColourDetectionTabModel
 
-from .components.combobox import Ui_Combobox
-from .components.label import Ui_Label
-from .components.line import Ui_Line
-from .components.pushbutton import Ui_PushButton
-from .components.radiobutton import Ui_RadioButton
-from .components.slider import Ui_Slider
+from ..components import (Ui_Combobox, Ui_Label, Ui_Line, Ui_PushButton,
+                          Ui_RadioButton, Ui_Slider)
 
 
 class ColourDetectionTabView(QtWidgets.QWidget):
