+++ src/snooker_ball_tracker/views/video_player/__init__.py
@@ -1,0 +1,2 @@
+from .options import Ui_VideoPlayerOptions
+from .player import Ui_VideoPlayer
